+++ lib/config/routes/routes.dart
@@ -12,22 +12,20 @@ final _routes = <RouteBase>[
   GoRoute(
     path: '/user',
     builder: (BuildContext context, GoRouterState state) {
-      return const UserPage();
+      return const UserHome(); // contain login page
     },
     routes: [
+      GoRoute(
+        path: 'login',
+        builder: (context, state) => const LoginPage(),
+      ),
       GoRoute(
         path: 'change-password',
         builder: (context, state) => const ChangePasswordPage(),
       ),
       GoRoute(
-        path: 'login',
-        builder: (context, state) => const LoginPage(),
-        routes: [
-          GoRoute(
-            path: 'forgot-password',
-            builder: (context, state) => const ForgotPasswordPage(),
-          ),
-        ],
+        path: 'forgot-password',
+        builder: (context, state) => const ForgotPasswordPage(),
       ),
     ],
   ),