+++ test/helpers/dummy_data/test_data.dart
@@ -1,0 +1,3 @@
+var testHeaders = {
+  'Content-Type': 'application/json',
+};